Add some Welsh translations
This commit adds some Welsh translations from our downstream project.
The list is only partial, as some strings do not occur in our project so
were not translated, or were sufficiently divorced from their original
meaning that to copy them across to this project would be inappropriate.

However, it does include the two legal boilerplate pages, which will
likely be useful for any subsequent projects using this codebase.

These messages were translated by the Department of Work and Pensions on
behalf of the Home Office.